Commit
Made the firebase data request functionality into a custom hook as one with do with Axios - useAxios or Fetch - useFetch. Test shows this is working. Integration of a loader and err message functions will take place withy a ternary operator.

Making a new design for the card components. Goal is to create a function to change the design of the card on click, as well as making a responsive design for smaller screen.

Continuing to look into making a dynamic route to get data on specific contact via the <Profile /> route.